requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2
FlagEmbedding==1.2.10
rank-bm25==0.2.2
pydantic==2.7.1
requests==2.32.3
streamlit==1.36.0
pydantic-settings==2.3.4
pytest==8.3.2
pytest-html==4.1.1
pytest-cov==5.0.0
